+++ gipit-react/app/actions/fetchDashboardStats.tsx
@@ -1,7 +1,7 @@
 'use server';
 
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/api/auth/authOptions";
 
 interface DashboardStats {
   activosCount: number;
