+++ gipit-react/app/actions/fetchInvoiceDetails.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/api/auth/authOptions";
 
 export const fetchInvoiceDetails = async (invoiceId: number) => {
   try {
