+++ gipit-react/app/actions/fetchInvoices.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/api/auth/authOptions";
 
 export const fetchAllPreInvoices = async (page: number) => {
   try {
