+++ gipit-react/next.config.mjs
@@ -1,8 +1,5 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    typescript: {
-        ignoreBuildErrors: true,
-     },
     images:{ remotePatterns: [
         {
             protocol: "http",
